add MultiBandReader and MultiBaseReader TilerFactories
